% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnyOfGameContentMediaEpgItems.r
\name{AnyOfGameContentMediaEpgItems}
\alias{AnyOfGameContentMediaEpgItems}
\title{AnyOfGameContentMediaEpgItems Class}
\description{
AnyOfGameContentMediaEpgItems Class

AnyOfGameContentMediaEpgItems Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{AnyOfGameContentMediaEpgItems$new()}}
\item \href{#method-toJSON}{\code{AnyOfGameContentMediaEpgItems$toJSON()}}
\item \href{#method-fromJSON}{\code{AnyOfGameContentMediaEpgItems$fromJSON()}}
\item \href{#method-toJSONString}{\code{AnyOfGameContentMediaEpgItems$toJSONString()}}
\item \href{#method-fromJSONString}{\code{AnyOfGameContentMediaEpgItems$fromJSONString()}}
\item \href{#method-clone}{\code{AnyOfGameContentMediaEpgItems$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$fromJSON(AnyOfGameContentMediaEpgItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$fromJSONString(AnyOfGameContentMediaEpgItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnyOfGameContentMediaEpgItems$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
