% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConferencesApi.r
\name{ConferencesApi}
\alias{ConferencesApi}
\title{Conferences operations}
\description{
swagger.Conferences
}
\section{Methods}{

\describe{

get_conference Get an NHL conference.


get_conferences Get all current NHL conferences.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ConferencesApi$new()}}
\item \href{#method-get_conference}{\code{ConferencesApi$get_conference()}}
\item \href{#method-get_conferences}{\code{ConferencesApi$get_conferences()}}
\item \href{#method-clone}{\code{ConferencesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferencesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_conference"></a>}}
\if{latex}{\out{\hypertarget{method-get_conference}{}}}
\subsection{Method \code{get_conference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferencesApi$get_conference(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_conferences"></a>}}
\if{latex}{\out{\hypertarget{method-get_conferences}{}}}
\subsection{Method \code{get_conferences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferencesApi$get_conferences(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConferencesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
