% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivisionsApi.r
\name{DivisionsApi}
\alias{DivisionsApi}
\title{Divisions operations}
\description{
swagger.Divisions
}
\section{Methods}{

\describe{

get_division Get an NHL division.


get_divisions Get all current NHL divisions.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DivisionsApi$new()}}
\item \href{#method-get_division}{\code{DivisionsApi$get_division()}}
\item \href{#method-get_divisions}{\code{DivisionsApi$get_divisions()}}
\item \href{#method-clone}{\code{DivisionsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DivisionsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_division"></a>}}
\if{latex}{\out{\hypertarget{method-get_division}{}}}
\subsection{Method \code{get_division()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DivisionsApi$get_division(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_divisions"></a>}}
\if{latex}{\out{\hypertarget{method-get_divisions}{}}}
\subsection{Method \code{get_divisions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DivisionsApi$get_divisions(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DivisionsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
