% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftApi.r
\name{DraftApi}
\alias{DraftApi}
\title{Draft operations}
\description{
swagger.Draft
}
\section{Methods}{

\describe{

get_draft Get round-by-round data for current year&#x27;s NHL Entry Draft.


get_draft_by_year Get round-by-round data for a specific year&#x27;s NHL Entry Draft.


get_draft_prospect Get an NHL Entry Draft prospect.


get_draft_prospects Get all NHL Entry Draft prospects.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DraftApi$new()}}
\item \href{#method-get_draft}{\code{DraftApi$get_draft()}}
\item \href{#method-get_draft_by_year}{\code{DraftApi$get_draft_by_year()}}
\item \href{#method-get_draft_prospect}{\code{DraftApi$get_draft_prospect()}}
\item \href{#method-get_draft_prospects}{\code{DraftApi$get_draft_prospects()}}
\item \href{#method-clone}{\code{DraftApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_draft"></a>}}
\if{latex}{\out{\hypertarget{method-get_draft}{}}}
\subsection{Method \code{get_draft()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$get_draft(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_draft_by_year"></a>}}
\if{latex}{\out{\hypertarget{method-get_draft_by_year}{}}}
\subsection{Method \code{get_draft_by_year()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$get_draft_by_year(year, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_draft_prospect"></a>}}
\if{latex}{\out{\hypertarget{method-get_draft_prospect}{}}}
\subsection{Method \code{get_draft_prospect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$get_draft_prospect(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_draft_prospects"></a>}}
\if{latex}{\out{\hypertarget{method-get_draft_prospects}{}}}
\subsection{Method \code{get_draft_prospects()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$get_draft_prospects(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
