% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftProspectPrimaryPosition.r
\name{DraftProspectPrimaryPosition}
\alias{DraftProspectPrimaryPosition}
\title{DraftProspectPrimaryPosition Class}
\description{
DraftProspectPrimaryPosition Class

DraftProspectPrimaryPosition Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{code}}{}

\item{\code{name}}{}

\item{\code{type}}{}

\item{\code{abbreviation}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{DraftProspectPrimaryPosition$new()}}
\item \href{#method-toJSON}{\code{DraftProspectPrimaryPosition$toJSON()}}
\item \href{#method-fromJSON}{\code{DraftProspectPrimaryPosition$fromJSON()}}
\item \href{#method-toJSONString}{\code{DraftProspectPrimaryPosition$toJSONString()}}
\item \href{#method-fromJSONString}{\code{DraftProspectPrimaryPosition$fromJSONString()}}
\item \href{#method-clone}{\code{DraftProspectPrimaryPosition$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$new(code, name, type, abbreviation)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$fromJSON(DraftProspectPrimaryPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$fromJSONString(DraftProspectPrimaryPositionJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DraftProspectPrimaryPosition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
