% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameBoxscoreTeam.r
\name{GameBoxscoreTeam}
\alias{GameBoxscoreTeam}
\title{GameBoxscoreTeam Class}
\description{
GameBoxscoreTeam Class

GameBoxscoreTeam Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{teamStats}}{}

\item{\code{players}}{}

\item{\code{goalies}}{}

\item{\code{skaters}}{}

\item{\code{onIce}}{}

\item{\code{onIcePlus}}{}

\item{\code{scratches}}{}

\item{\code{penaltyBox}}{}

\item{\code{coaches}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameBoxscoreTeam$new()}}
\item \href{#method-toJSON}{\code{GameBoxscoreTeam$toJSON()}}
\item \href{#method-fromJSON}{\code{GameBoxscoreTeam$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameBoxscoreTeam$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameBoxscoreTeam$fromJSONString()}}
\item \href{#method-clone}{\code{GameBoxscoreTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$new(
  team,
  teamStats,
  players,
  goalies,
  skaters,
  onIce,
  onIcePlus,
  scratches,
  penaltyBox,
  coaches
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$fromJSON(GameBoxscoreTeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$fromJSONString(GameBoxscoreTeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
