% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameBoxscoreTeamOnIcePlus.r
\name{GameBoxscoreTeamOnIcePlus}
\alias{GameBoxscoreTeamOnIcePlus}
\title{GameBoxscoreTeamOnIcePlus Class}
\description{
GameBoxscoreTeamOnIcePlus Class

GameBoxscoreTeamOnIcePlus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{playerId}}{}

\item{\code{shiftDuration}}{}

\item{\code{stamina}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameBoxscoreTeamOnIcePlus$new()}}
\item \href{#method-toJSON}{\code{GameBoxscoreTeamOnIcePlus$toJSON()}}
\item \href{#method-fromJSON}{\code{GameBoxscoreTeamOnIcePlus$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameBoxscoreTeamOnIcePlus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameBoxscoreTeamOnIcePlus$fromJSONString()}}
\item \href{#method-clone}{\code{GameBoxscoreTeamOnIcePlus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$new(playerId, shiftDuration, stamina)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$fromJSON(GameBoxscoreTeamOnIcePlusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$fromJSONString(GameBoxscoreTeamOnIcePlusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamOnIcePlus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
