% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameBoxscoreTeamPerson.r
\name{GameBoxscoreTeamPerson}
\alias{GameBoxscoreTeamPerson}
\title{GameBoxscoreTeamPerson Class}
\description{
GameBoxscoreTeamPerson Class

GameBoxscoreTeamPerson Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{fullName}}{}

\item{\code{link}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameBoxscoreTeamPerson$new()}}
\item \href{#method-toJSON}{\code{GameBoxscoreTeamPerson$toJSON()}}
\item \href{#method-fromJSON}{\code{GameBoxscoreTeamPerson$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameBoxscoreTeamPerson$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameBoxscoreTeamPerson$fromJSONString()}}
\item \href{#method-clone}{\code{GameBoxscoreTeamPerson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$new(fullName, link)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$fromJSON(GameBoxscoreTeamPersonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$fromJSONString(GameBoxscoreTeamPersonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPerson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
