% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameBoxscoreTeamPlayersPerson.r
\name{GameBoxscoreTeamPlayersPerson}
\alias{GameBoxscoreTeamPlayersPerson}
\title{GameBoxscoreTeamPlayersPerson Class}
\description{
GameBoxscoreTeamPlayersPerson Class

GameBoxscoreTeamPlayersPerson Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{fullName}}{}

\item{\code{link}}{}

\item{\code{shootsCatches}}{}

\item{\code{rosterStatus}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameBoxscoreTeamPlayersPerson$new()}}
\item \href{#method-toJSON}{\code{GameBoxscoreTeamPlayersPerson$toJSON()}}
\item \href{#method-fromJSON}{\code{GameBoxscoreTeamPlayersPerson$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameBoxscoreTeamPlayersPerson$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameBoxscoreTeamPlayersPerson$fromJSONString()}}
\item \href{#method-clone}{\code{GameBoxscoreTeamPlayersPerson$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$new(
  id,
  fullName,
  link,
  shootsCatches,
  rosterStatus
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$fromJSON(GameBoxscoreTeamPlayersPersonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$fromJSONString(GameBoxscoreTeamPlayersPersonJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersPerson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
