% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameBoxscoreTeamPlayersStatsSkaterStats.r
\name{GameBoxscoreTeamPlayersStatsSkaterStats}
\alias{GameBoxscoreTeamPlayersStatsSkaterStats}
\title{GameBoxscoreTeamPlayersStatsSkaterStats Class}
\description{
GameBoxscoreTeamPlayersStatsSkaterStats Class

GameBoxscoreTeamPlayersStatsSkaterStats Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeOnIce}}{}

\item{\code{assists}}{}

\item{\code{goals}}{}

\item{\code{shots}}{}

\item{\code{hits}}{}

\item{\code{powerPlayGoals}}{}

\item{\code{powerPlayAssists}}{}

\item{\code{penaltyMinutes}}{}

\item{\code{faceOffWins}}{}

\item{\code{faceoffTaken}}{}

\item{\code{takeaways}}{}

\item{\code{giveaways}}{}

\item{\code{shortHandedGoals}}{}

\item{\code{shortHandedAssists}}{}

\item{\code{blocked}}{}

\item{\code{plusMinus}}{}

\item{\code{evenTimeOnIce}}{}

\item{\code{powerPlayTimeOnIce}}{}

\item{\code{shortHandedTimeOnIce}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$new()}}
\item \href{#method-toJSON}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$toJSON()}}
\item \href{#method-fromJSON}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$fromJSONString()}}
\item \href{#method-clone}{\code{GameBoxscoreTeamPlayersStatsSkaterStats$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$new(
  timeOnIce,
  assists,
  goals,
  shots,
  hits,
  powerPlayGoals,
  powerPlayAssists,
  penaltyMinutes,
  faceOffWins,
  faceoffTaken,
  takeaways,
  giveaways,
  shortHandedGoals,
  shortHandedAssists,
  blocked,
  plusMinus,
  evenTimeOnIce,
  powerPlayTimeOnIce,
  shortHandedTimeOnIce
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$fromJSON(
  GameBoxscoreTeamPlayersStatsSkaterStatsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$fromJSONString(
  GameBoxscoreTeamPlayersStatsSkaterStatsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameBoxscoreTeamPlayersStatsSkaterStats$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
