% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameContentMediaMilestonesItems.r
\name{GameContentMediaMilestonesItems}
\alias{GameContentMediaMilestonesItems}
\title{GameContentMediaMilestonesItems Class}
\description{
GameContentMediaMilestonesItems Class

GameContentMediaMilestonesItems Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{title}}{}

\item{\code{description}}{}

\item{\code{type}}{}

\item{\code{timeAbsolute}}{}

\item{\code{timeOffset}}{}

\item{\code{period}}{}

\item{\code{statsEventId}}{}

\item{\code{teamId}}{}

\item{\code{playerId}}{}

\item{\code{periodTime}}{}

\item{\code{ordinalNum}}{}

\item{\code{highlight}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameContentMediaMilestonesItems$new()}}
\item \href{#method-toJSON}{\code{GameContentMediaMilestonesItems$toJSON()}}
\item \href{#method-fromJSON}{\code{GameContentMediaMilestonesItems$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameContentMediaMilestonesItems$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameContentMediaMilestonesItems$fromJSONString()}}
\item \href{#method-clone}{\code{GameContentMediaMilestonesItems$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$new(
  title,
  description,
  type,
  timeAbsolute,
  timeOffset,
  period,
  statsEventId,
  teamId,
  playerId,
  periodTime,
  ordinalNum,
  highlight
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$fromJSON(GameContentMediaMilestonesItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$fromJSONString(
  GameContentMediaMilestonesItemsJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameContentMediaMilestonesItems$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
