% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameEditorial.r
\name{GameEditorial}
\alias{GameEditorial}
\title{GameEditorial Class}
\description{
GameEditorial Class

GameEditorial Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{}

\item{\code{state}}{}

\item{\code{date}}{}

\item{\code{id}}{}

\item{\code{headline}}{}

\item{\code{subhead}}{}

\item{\code{seoTitle}}{}

\item{\code{seoDescription}}{}

\item{\code{seoKeywords}}{}

\item{\code{slug}}{}

\item{\code{commenting}}{}

\item{\code{tagline}}{}

\item{\code{tokenData}}{}

\item{\code{contributor}}{}

\item{\code{keywordsDisplay}}{}

\item{\code{keywordsAll}}{}

\item{\code{approval}}{}

\item{\code{url}}{}

\item{\code{dataURI}}{}

\item{\code{primaryKeyword}}{}

\item{\code{media}}{}

\item{\code{preview}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameEditorial$new()}}
\item \href{#method-toJSON}{\code{GameEditorial$toJSON()}}
\item \href{#method-fromJSON}{\code{GameEditorial$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameEditorial$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameEditorial$fromJSONString()}}
\item \href{#method-clone}{\code{GameEditorial$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$new(
  type,
  state,
  date,
  id,
  headline,
  subhead,
  seoTitle,
  seoDescription,
  seoKeywords,
  slug,
  commenting,
  tagline,
  tokenData,
  contributor,
  keywordsDisplay,
  keywordsAll,
  approval,
  url,
  dataURI,
  primaryKeyword,
  media,
  preview
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$fromJSON(GameEditorialJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$fromJSONString(GameEditorialJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorial$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
