% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameEditorialTokenData.r
\name{GameEditorialTokenData}
\alias{GameEditorialTokenData}
\title{GameEditorialTokenData Class}
\description{
GameEditorialTokenData Class

GameEditorialTokenData Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tokenGUID}}{}

\item{\code{type}}{}

\item{\code{id}}{}

\item{\code{teamId}}{}

\item{\code{name}}{}

\item{\code{seoName}}{}

\item{\code{href}}{}

\item{\code{hrefMobile}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameEditorialTokenData$new()}}
\item \href{#method-toJSON}{\code{GameEditorialTokenData$toJSON()}}
\item \href{#method-fromJSON}{\code{GameEditorialTokenData$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameEditorialTokenData$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameEditorialTokenData$fromJSONString()}}
\item \href{#method-clone}{\code{GameEditorialTokenData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$new(
  tokenGUID,
  type,
  id,
  teamId,
  name,
  seoName,
  href,
  hrefMobile
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$fromJSON(GameEditorialTokenDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$fromJSONString(GameEditorialTokenDataJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameEditorialTokenData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
