% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameHighlight.r
\name{GameHighlight}
\alias{GameHighlight}
\title{GameHighlight Class}
\description{
GameHighlight Class

GameHighlight Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{}

\item{\code{id}}{}

\item{\code{date}}{}

\item{\code{title}}{}

\item{\code{blurb}}{}

\item{\code{description}}{}

\item{\code{duration}}{}

\item{\code{authFlow}}{}

\item{\code{mediaPlaybackId}}{}

\item{\code{mediaState}}{}

\item{\code{keywords}}{}

\item{\code{image}}{}

\item{\code{playbacks}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameHighlight$new()}}
\item \href{#method-toJSON}{\code{GameHighlight$toJSON()}}
\item \href{#method-fromJSON}{\code{GameHighlight$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameHighlight$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameHighlight$fromJSONString()}}
\item \href{#method-clone}{\code{GameHighlight$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$new(
  type,
  id,
  date,
  title,
  blurb,
  description,
  duration,
  authFlow,
  mediaPlaybackId,
  mediaState,
  keywords,
  image,
  playbacks
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$fromJSON(GameHighlightJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$fromJSONString(GameHighlightJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameHighlight$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
