% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLinescore.r
\name{GameLinescore}
\alias{GameLinescore}
\title{GameLinescore Class}
\description{
GameLinescore Class

GameLinescore Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{currentPeriod}}{}

\item{\code{currentPeriodOrdinal}}{}

\item{\code{currentPeriodTimeRemaining}}{}

\item{\code{periods}}{}

\item{\code{shootoutInfo}}{}

\item{\code{teams}}{}

\item{\code{powerPlayStrength}}{}

\item{\code{hasShootout}}{}

\item{\code{intermissionInfo}}{}

\item{\code{powerPlayInfo}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLinescore$new()}}
\item \href{#method-toJSON}{\code{GameLinescore$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLinescore$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLinescore$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLinescore$fromJSONString()}}
\item \href{#method-clone}{\code{GameLinescore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$new(
  currentPeriod,
  currentPeriodOrdinal,
  currentPeriodTimeRemaining,
  periods,
  shootoutInfo,
  teams,
  powerPlayStrength,
  hasShootout,
  intermissionInfo,
  powerPlayInfo
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$fromJSON(GameLinescoreJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$fromJSONString(GameLinescoreJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
