% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLinescorePowerPlayInfo.r
\name{GameLinescorePowerPlayInfo}
\alias{GameLinescorePowerPlayInfo}
\title{GameLinescorePowerPlayInfo Class}
\description{
GameLinescorePowerPlayInfo Class

GameLinescorePowerPlayInfo Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{situationTimeRemaining}}{}

\item{\code{situationTimeElapsed}}{}

\item{\code{inSituation}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLinescorePowerPlayInfo$new()}}
\item \href{#method-toJSON}{\code{GameLinescorePowerPlayInfo$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLinescorePowerPlayInfo$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLinescorePowerPlayInfo$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLinescorePowerPlayInfo$fromJSONString()}}
\item \href{#method-clone}{\code{GameLinescorePowerPlayInfo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$new(
  situationTimeRemaining,
  situationTimeElapsed,
  inSituation
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$fromJSON(GameLinescorePowerPlayInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$fromJSONString(GameLinescorePowerPlayInfoJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescorePowerPlayInfo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
