% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLinescoreShootoutInfoAway.r
\name{GameLinescoreShootoutInfoAway}
\alias{GameLinescoreShootoutInfoAway}
\title{GameLinescoreShootoutInfoAway Class}
\description{
GameLinescoreShootoutInfoAway Class

GameLinescoreShootoutInfoAway Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{scores}}{}

\item{\code{attempts}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLinescoreShootoutInfoAway$new()}}
\item \href{#method-toJSON}{\code{GameLinescoreShootoutInfoAway$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLinescoreShootoutInfoAway$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLinescoreShootoutInfoAway$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLinescoreShootoutInfoAway$fromJSONString()}}
\item \href{#method-clone}{\code{GameLinescoreShootoutInfoAway$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$new(scores, attempts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$fromJSON(GameLinescoreShootoutInfoAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$fromJSONString(GameLinescoreShootoutInfoAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreShootoutInfoAway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
