% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLinescoreTeam.r
\name{GameLinescoreTeam}
\alias{GameLinescoreTeam}
\title{GameLinescoreTeam Class}
\description{
GameLinescoreTeam Class

GameLinescoreTeam Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{goals}}{}

\item{\code{shotsOnGoal}}{}

\item{\code{goaliePulled}}{}

\item{\code{numSkaters}}{}

\item{\code{powerPlay}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLinescoreTeam$new()}}
\item \href{#method-toJSON}{\code{GameLinescoreTeam$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLinescoreTeam$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLinescoreTeam$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLinescoreTeam$fromJSONString()}}
\item \href{#method-clone}{\code{GameLinescoreTeam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$new(
  team,
  goals,
  shotsOnGoal,
  goaliePulled,
  numSkaters,
  powerPlay
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$fromJSON(GameLinescoreTeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$fromJSONString(GameLinescoreTeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLinescoreTeam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
