% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLiveDataPlays.r
\name{GameLiveDataPlays}
\alias{GameLiveDataPlays}
\title{GameLiveDataPlays Class}
\description{
GameLiveDataPlays Class

GameLiveDataPlays Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{allPlays}}{}

\item{\code{scoringPlays}}{}

\item{\code{penaltyPlays}}{}

\item{\code{playsByPeriod}}{}

\item{\code{currentPlay}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLiveDataPlays$new()}}
\item \href{#method-toJSON}{\code{GameLiveDataPlays$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLiveDataPlays$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLiveDataPlays$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLiveDataPlays$fromJSONString()}}
\item \href{#method-clone}{\code{GameLiveDataPlays$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$new(
  allPlays,
  scoringPlays,
  penaltyPlays,
  playsByPeriod,
  currentPlay
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$fromJSON(GameLiveDataPlaysJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$fromJSONString(GameLiveDataPlaysJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlays$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
