% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameLiveDataPlaysPlaysByPeriod.r
\name{GameLiveDataPlaysPlaysByPeriod}
\alias{GameLiveDataPlaysPlaysByPeriod}
\title{GameLiveDataPlaysPlaysByPeriod Class}
\description{
GameLiveDataPlaysPlaysByPeriod Class

GameLiveDataPlaysPlaysByPeriod Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{startIndex}}{}

\item{\code{plays}}{}

\item{\code{endIndex}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameLiveDataPlaysPlaysByPeriod$new()}}
\item \href{#method-toJSON}{\code{GameLiveDataPlaysPlaysByPeriod$toJSON()}}
\item \href{#method-fromJSON}{\code{GameLiveDataPlaysPlaysByPeriod$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameLiveDataPlaysPlaysByPeriod$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameLiveDataPlaysPlaysByPeriod$fromJSONString()}}
\item \href{#method-clone}{\code{GameLiveDataPlaysPlaysByPeriod$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$new(startIndex, plays, endIndex)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$fromJSON(GameLiveDataPlaysPlaysByPeriodJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$fromJSONString(
  GameLiveDataPlaysPlaysByPeriodJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameLiveDataPlaysPlaysByPeriod$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
