% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameMediaAudioItems.r
\name{GameMediaAudioItems}
\alias{GameMediaAudioItems}
\title{GameMediaAudioItems Class}
\description{
GameMediaAudioItems Class

GameMediaAudioItems Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mediaState}}{}

\item{\code{mediaPlaybackId}}{}

\item{\code{mediaFeedType}}{}

\item{\code{callLetters}}{}

\item{\code{eventId}}{}

\item{\code{language}}{}

\item{\code{freeGame}}{}

\item{\code{feedName}}{}

\item{\code{gamePlus}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameMediaAudioItems$new()}}
\item \href{#method-toJSON}{\code{GameMediaAudioItems$toJSON()}}
\item \href{#method-fromJSON}{\code{GameMediaAudioItems$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameMediaAudioItems$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameMediaAudioItems$fromJSONString()}}
\item \href{#method-clone}{\code{GameMediaAudioItems$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$new(
  mediaState,
  mediaPlaybackId,
  mediaFeedType,
  callLetters,
  eventId,
  language,
  freeGame,
  feedName,
  gamePlus
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$fromJSON(GameMediaAudioItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$fromJSONString(GameMediaAudioItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaAudioItems$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
