% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GameMediaNHLTVItems.r
\name{GameMediaNHLTVItems}
\alias{GameMediaNHLTVItems}
\title{GameMediaNHLTVItems Class}
\description{
GameMediaNHLTVItems Class

GameMediaNHLTVItems Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{guid}}{}

\item{\code{mediaState}}{}

\item{\code{mediaPlaybackId}}{}

\item{\code{mediaFeedType}}{}

\item{\code{callLetters}}{}

\item{\code{eventId}}{}

\item{\code{language}}{}

\item{\code{freeGame}}{}

\item{\code{feedName}}{}

\item{\code{gamePlus}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GameMediaNHLTVItems$new()}}
\item \href{#method-toJSON}{\code{GameMediaNHLTVItems$toJSON()}}
\item \href{#method-fromJSON}{\code{GameMediaNHLTVItems$fromJSON()}}
\item \href{#method-toJSONString}{\code{GameMediaNHLTVItems$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GameMediaNHLTVItems$fromJSONString()}}
\item \href{#method-clone}{\code{GameMediaNHLTVItems$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$new(
  guid,
  mediaState,
  mediaPlaybackId,
  mediaFeedType,
  callLetters,
  eventId,
  language,
  freeGame,
  feedName,
  gamePlus
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$fromJSON(GameMediaNHLTVItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$fromJSONString(GameMediaNHLTVItemsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GameMediaNHLTVItems$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
