% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamePlayAbout.r
\name{GamePlayAbout}
\alias{GamePlayAbout}
\title{GamePlayAbout Class}
\description{
GamePlayAbout Class

GamePlayAbout Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{eventIdx}}{}

\item{\code{eventId}}{}

\item{\code{period}}{}

\item{\code{periodType}}{}

\item{\code{ordinalNum}}{}

\item{\code{periodTime}}{}

\item{\code{periodTimeRemaining}}{}

\item{\code{dateTime}}{}

\item{\code{goals}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GamePlayAbout$new()}}
\item \href{#method-toJSON}{\code{GamePlayAbout$toJSON()}}
\item \href{#method-fromJSON}{\code{GamePlayAbout$fromJSON()}}
\item \href{#method-toJSONString}{\code{GamePlayAbout$toJSONString()}}
\item \href{#method-fromJSONString}{\code{GamePlayAbout$fromJSONString()}}
\item \href{#method-clone}{\code{GamePlayAbout$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$new(
  eventIdx,
  eventId,
  period,
  periodType,
  ordinalNum,
  periodTime,
  periodTimeRemaining,
  dateTime,
  goals
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$fromJSON(GamePlayAboutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$fromJSONString(GamePlayAboutJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamePlayAbout$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
