% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GamesApi.r
\name{GamesApi}
\alias{GamesApi}
\title{Games operations}
\description{
swagger.Games
}
\section{Methods}{

\describe{

get_game Get all available data for an NHL game.


get_game_boxscore Get the boxscore for an NHL game.


get_game_content Get editorials, video replays and photo highlights for an NHL game.


get_game_diff Get all available data for an NHL game after a specific time.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{GamesApi$new()}}
\item \href{#method-get_game}{\code{GamesApi$get_game()}}
\item \href{#method-get_game_boxscore}{\code{GamesApi$get_game_boxscore()}}
\item \href{#method-get_game_content}{\code{GamesApi$get_game_content()}}
\item \href{#method-get_game_diff}{\code{GamesApi$get_game_diff()}}
\item \href{#method-clone}{\code{GamesApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game"></a>}}
\if{latex}{\out{\hypertarget{method-get_game}{}}}
\subsection{Method \code{get_game()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_game(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game_boxscore"></a>}}
\if{latex}{\out{\hypertarget{method-get_game_boxscore}{}}}
\subsection{Method \code{get_game_boxscore()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_game_boxscore(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game_content"></a>}}
\if{latex}{\out{\hypertarget{method-get_game_content}{}}}
\subsection{Method \code{get_game_content()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_game_content(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_game_diff"></a>}}
\if{latex}{\out{\hypertarget{method-get_game_diff}{}}}
\subsection{Method \code{get_game_diff()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$get_game_diff(id, start_time_code, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GamesApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
