% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_player_stats.R
\name{nhl_player_stats}
\alias{nhl_player_stats}
\title{NHL Player Stats}
\usage{
nhl_player_stats(player_id)
}
\arguments{
\item{player_id}{Player unique ID}
}
\value{
Returns a named list of data frames: player, stats
}
\description{
Returns player stats for a given player ID
}
\examples{
\donttest{
  nhl_player_stats(player_id=8476899)
}
}
\keyword{NHL}
\keyword{Player}
\keyword{stats}
