% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_schedule.R
\name{nhl_schedule}
\alias{nhl_schedule}
\title{NHL Schedule}
\usage{
nhl_schedule(season = NULL, day = as.Date(Sys.Date(), "\%Y-\%m-\%d"))
}
\arguments{
\item{season}{NHL Season}

\item{day}{Date}
}
\value{
Returns a named list of data frames: player, stats
}
\description{
Returns NHL Schedule data
}
\examples{
\donttest{
  nhl_schedule(season = 2020)
}
}
\keyword{NHL}
\keyword{Player}
\keyword{stats}
