% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Player.r
\name{Player}
\alias{Player}
\title{Player Class}
\description{
Player Class

Player Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{fullName}}{}

\item{\code{link}}{}

\item{\code{firstName}}{}

\item{\code{lastName}}{}

\item{\code{primaryNumber}}{}

\item{\code{birthDate}}{}

\item{\code{currentAge}}{}

\item{\code{birthCity}}{}

\item{\code{birthStateProvince}}{}

\item{\code{birthCountry}}{}

\item{\code{nationality}}{}

\item{\code{height}}{}

\item{\code{weight}}{}

\item{\code{active}}{}

\item{\code{alternateCaptain}}{}

\item{\code{captain}}{}

\item{\code{rookie}}{}

\item{\code{shootsCatches}}{}

\item{\code{rosterStatus}}{}

\item{\code{currentTeam}}{}

\item{\code{primaryPosition}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Player$new()}}
\item \href{#method-toJSON}{\code{Player$toJSON()}}
\item \href{#method-fromJSON}{\code{Player$fromJSON()}}
\item \href{#method-toJSONString}{\code{Player$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Player$fromJSONString()}}
\item \href{#method-clone}{\code{Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$new(
  id,
  fullName,
  link,
  firstName,
  lastName,
  primaryNumber,
  birthDate,
  currentAge,
  birthCity,
  birthStateProvince,
  birthCountry,
  nationality,
  height,
  weight,
  active,
  alternateCaptain,
  captain,
  rookie,
  shootsCatches,
  rosterStatus,
  currentTeam,
  primaryPosition
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$fromJSON(PlayerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$fromJSONString(PlayerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
