% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayersApi.r
\name{PlayersApi}
\alias{PlayersApi}
\title{Players operations}
\description{
swagger.Players
}
\section{Methods}{

\describe{

get_player Get an NHL player.


get_player_stats Get specific statistics for an NHL player.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayersApi$new()}}
\item \href{#method-get_player}{\code{PlayersApi$get_player()}}
\item \href{#method-get_player_stats}{\code{PlayersApi$get_player_stats()}}
\item \href{#method-clone}{\code{PlayersApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayersApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player"></a>}}
\if{latex}{\out{\hypertarget{method-get_player}{}}}
\subsection{Method \code{get_player()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayersApi$get_player(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_player_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_player_stats}{}}}
\subsection{Method \code{get_player_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayersApi$get_player_stats(id, stats, season, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayersApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
