% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlayerStatsStat.r
\name{PlayerStatsStat}
\alias{PlayerStatsStat}
\title{PlayerStatsStat Class}
\description{
PlayerStatsStat Class

PlayerStatsStat Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{timeOnIce}}{}

\item{\code{assists}}{}

\item{\code{goals}}{}

\item{\code{pim}}{}

\item{\code{shots}}{}

\item{\code{games}}{}

\item{\code{hits}}{}

\item{\code{powerPlayGoals}}{}

\item{\code{powerPlayPoints}}{}

\item{\code{powerPlayTimeOnIce}}{}

\item{\code{evenTimeOnIce}}{}

\item{\code{penaltyMinutes}}{}

\item{\code{faceOffPct}}{}

\item{\code{shotPct}}{}

\item{\code{gameWinningGoals}}{}

\item{\code{overTimeGoals}}{}

\item{\code{shortHandedGoals}}{}

\item{\code{shortHandedPoints}}{}

\item{\code{shortHandedTimeOnIce}}{}

\item{\code{blocked}}{}

\item{\code{plusMinus}}{}

\item{\code{points}}{}

\item{\code{shifts}}{}

\item{\code{timeOnIcePerGame}}{}

\item{\code{evenTimeOnIcePerGame}}{}

\item{\code{shortHandedTimeOnIcePerGame}}{}

\item{\code{powerPlayTimeOnIcePerGame}}{}

\item{\code{rankPowerPlayGoals}}{}

\item{\code{rankBlockedShots}}{}

\item{\code{rankAssists}}{}

\item{\code{rankShotPct}}{}

\item{\code{rankGoals}}{}

\item{\code{rankHits}}{}

\item{\code{rankPenaltyMinutes}}{}

\item{\code{rankShortHandedGoals}}{}

\item{\code{rankPlusMinus}}{}

\item{\code{rankShots}}{}

\item{\code{rankPoints}}{}

\item{\code{rankOvertimeGoals}}{}

\item{\code{rankGamesPlayed}}{}

\item{\code{goalsInFirstPeriod}}{}

\item{\code{goalsInSecondPeriod}}{}

\item{\code{goalsInThirdPeriod}}{}

\item{\code{goalsTrailingByOne}}{}

\item{\code{goalsTrailingByTwo}}{}

\item{\code{goalsTrailingByThreePlus}}{}

\item{\code{goalsWhenTied}}{}

\item{\code{goalsLeadingByOne}}{}

\item{\code{goalsLeadingByTwo}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PlayerStatsStat$new()}}
\item \href{#method-toJSON}{\code{PlayerStatsStat$toJSON()}}
\item \href{#method-fromJSON}{\code{PlayerStatsStat$fromJSON()}}
\item \href{#method-toJSONString}{\code{PlayerStatsStat$toJSONString()}}
\item \href{#method-fromJSONString}{\code{PlayerStatsStat$fromJSONString()}}
\item \href{#method-clone}{\code{PlayerStatsStat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$new(
  timeOnIce,
  assists,
  goals,
  pim,
  shots,
  games,
  hits,
  powerPlayGoals,
  powerPlayPoints,
  powerPlayTimeOnIce,
  evenTimeOnIce,
  penaltyMinutes,
  faceOffPct,
  shotPct,
  gameWinningGoals,
  overTimeGoals,
  shortHandedGoals,
  shortHandedPoints,
  shortHandedTimeOnIce,
  blocked,
  plusMinus,
  points,
  shifts,
  timeOnIcePerGame,
  evenTimeOnIcePerGame,
  shortHandedTimeOnIcePerGame,
  powerPlayTimeOnIcePerGame,
  rankPowerPlayGoals,
  rankBlockedShots,
  rankAssists,
  rankShotPct,
  rankGoals,
  rankHits,
  rankPenaltyMinutes,
  rankShortHandedGoals,
  rankPlusMinus,
  rankShots,
  rankPoints,
  rankOvertimeGoals,
  rankGamesPlayed,
  goalsInFirstPeriod,
  goalsInSecondPeriod,
  goalsInThirdPeriod,
  goalsTrailingByOne,
  goalsTrailingByTwo,
  goalsTrailingByThreePlus,
  goalsWhenTied,
  goalsLeadingByOne,
  goalsLeadingByTwo
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$fromJSON(PlayerStatsStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$fromJSONString(PlayerStatsStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlayerStatsStat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
