% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduleApi.r
\name{ScheduleApi}
\alias{ScheduleApi}
\title{Schedule operations}
\description{
swagger.Schedule
}
\section{Methods}{

\describe{

get_schedule Get the NHL game schedule.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduleApi$new()}}
\item \href{#method-get_schedule}{\code{ScheduleApi$get_schedule()}}
\item \href{#method-clone}{\code{ScheduleApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-get_schedule}{}}}
\subsection{Method \code{get_schedule()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleApi$get_schedule(expand, team_id, start_date, end_date, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
