% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduleGame.r
\name{ScheduleGame}
\alias{ScheduleGame}
\title{ScheduleGame Class}
\description{
ScheduleGame Class

ScheduleGame Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gamePk}}{}

\item{\code{link}}{}

\item{\code{gameType}}{}

\item{\code{season}}{}

\item{\code{gameDate}}{}

\item{\code{status}}{}

\item{\code{teams}}{}

\item{\code{linescore}}{}

\item{\code{venue}}{}

\item{\code{tickets}}{}

\item{\code{content}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduleGame$new()}}
\item \href{#method-toJSON}{\code{ScheduleGame$toJSON()}}
\item \href{#method-fromJSON}{\code{ScheduleGame$fromJSON()}}
\item \href{#method-toJSONString}{\code{ScheduleGame$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScheduleGame$fromJSONString()}}
\item \href{#method-clone}{\code{ScheduleGame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$new(
  gamePk,
  link,
  gameType,
  season,
  gameDate,
  status,
  teams,
  linescore,
  venue,
  tickets,
  content
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$fromJSON(ScheduleGameJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$fromJSONString(ScheduleGameJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
