% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduleGameTeamsAway.r
\name{ScheduleGameTeamsAway}
\alias{ScheduleGameTeamsAway}
\title{ScheduleGameTeamsAway Class}
\description{
ScheduleGameTeamsAway Class

ScheduleGameTeamsAway Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{leagueRecord}}{}

\item{\code{score}}{}

\item{\code{team}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduleGameTeamsAway$new()}}
\item \href{#method-toJSON}{\code{ScheduleGameTeamsAway$toJSON()}}
\item \href{#method-fromJSON}{\code{ScheduleGameTeamsAway$fromJSON()}}
\item \href{#method-toJSONString}{\code{ScheduleGameTeamsAway$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScheduleGameTeamsAway$fromJSONString()}}
\item \href{#method-clone}{\code{ScheduleGameTeamsAway$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$new(leagueRecord, score, team)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$fromJSON(ScheduleGameTeamsAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$fromJSONString(ScheduleGameTeamsAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
