% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScheduleGameTeamsAwayLeagueRecord.r
\name{ScheduleGameTeamsAwayLeagueRecord}
\alias{ScheduleGameTeamsAwayLeagueRecord}
\title{ScheduleGameTeamsAwayLeagueRecord Class}
\description{
ScheduleGameTeamsAwayLeagueRecord Class

ScheduleGameTeamsAwayLeagueRecord Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{wins}}{}

\item{\code{losses}}{}

\item{\code{ot}}{}

\item{\code{type}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ScheduleGameTeamsAwayLeagueRecord$new()}}
\item \href{#method-toJSON}{\code{ScheduleGameTeamsAwayLeagueRecord$toJSON()}}
\item \href{#method-fromJSON}{\code{ScheduleGameTeamsAwayLeagueRecord$fromJSON()}}
\item \href{#method-toJSONString}{\code{ScheduleGameTeamsAwayLeagueRecord$toJSONString()}}
\item \href{#method-fromJSONString}{\code{ScheduleGameTeamsAwayLeagueRecord$fromJSONString()}}
\item \href{#method-clone}{\code{ScheduleGameTeamsAwayLeagueRecord$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$new(wins, losses, ot, type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$fromJSON(
  ScheduleGameTeamsAwayLeagueRecordJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$fromJSONString(
  ScheduleGameTeamsAwayLeagueRecordJson
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScheduleGameTeamsAwayLeagueRecord$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
