% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandingsApi.r
\name{StandingsApi}
\alias{StandingsApi}
\title{Standings operations}
\description{
swagger.Standings
}
\section{Methods}{

\describe{

get_standing_types Get all available NHL standing types.


get_standings Get NHL division standings.


get_standings_by_type Get NHL standings for a specific standing type.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StandingsApi$new()}}
\item \href{#method-get_standing_types}{\code{StandingsApi$get_standing_types()}}
\item \href{#method-get_standings}{\code{StandingsApi$get_standings()}}
\item \href{#method-get_standings_by_type}{\code{StandingsApi$get_standings_by_type()}}
\item \href{#method-clone}{\code{StandingsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_standing_types"></a>}}
\if{latex}{\out{\hypertarget{method-get_standing_types}{}}}
\subsection{Method \code{get_standing_types()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsApi$get_standing_types(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_standings"></a>}}
\if{latex}{\out{\hypertarget{method-get_standings}{}}}
\subsection{Method \code{get_standings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsApi$get_standings(season, date, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_standings_by_type"></a>}}
\if{latex}{\out{\hypertarget{method-get_standings_by_type}{}}}
\subsection{Method \code{get_standings_by_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsApi$get_standings_by_type(type, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
