% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandingsTeamRecords.r
\name{StandingsTeamRecords}
\alias{StandingsTeamRecords}
\title{StandingsTeamRecords Class}
\description{
StandingsTeamRecords Class

StandingsTeamRecords Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{team}}{}

\item{\code{leagueRecord}}{}

\item{\code{goalsAgainst}}{}

\item{\code{goalsScored}}{}

\item{\code{points}}{}

\item{\code{divisionRank}}{}

\item{\code{conferenceRank}}{}

\item{\code{leagueRank}}{}

\item{\code{wildCardRank}}{}

\item{\code{row}}{}

\item{\code{gamesPlayed}}{}

\item{\code{streak}}{}

\item{\code{lastUpdated}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{StandingsTeamRecords$new()}}
\item \href{#method-toJSON}{\code{StandingsTeamRecords$toJSON()}}
\item \href{#method-fromJSON}{\code{StandingsTeamRecords$fromJSON()}}
\item \href{#method-toJSONString}{\code{StandingsTeamRecords$toJSONString()}}
\item \href{#method-fromJSONString}{\code{StandingsTeamRecords$fromJSONString()}}
\item \href{#method-clone}{\code{StandingsTeamRecords$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$new(
  team,
  leagueRecord,
  goalsAgainst,
  goalsScored,
  points,
  divisionRank,
  conferenceRank,
  leagueRank,
  wildCardRank,
  row,
  gamesPlayed,
  streak,
  lastUpdated
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$fromJSON(StandingsTeamRecordsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$fromJSONString(StandingsTeamRecordsJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{StandingsTeamRecords$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
