% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Team.r
\name{Team}
\alias{Team}
\title{Team Class}
\description{
Team Class

Team Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{}

\item{\code{name}}{}

\item{\code{link}}{}

\item{\code{venue}}{}

\item{\code{abbreviation}}{}

\item{\code{triCode}}{}

\item{\code{teamName}}{}

\item{\code{locationName}}{}

\item{\code{firstYearOfPlay}}{}

\item{\code{division}}{}

\item{\code{conference}}{}

\item{\code{franchise}}{}

\item{\code{roster}}{}

\item{\code{nextGameSchedule}}{}

\item{\code{shortName}}{}

\item{\code{officialSiteUrl}}{}

\item{\code{franchiseId}}{}

\item{\code{active}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Team$new()}}
\item \href{#method-toJSON}{\code{Team$toJSON()}}
\item \href{#method-fromJSON}{\code{Team$fromJSON()}}
\item \href{#method-toJSONString}{\code{Team$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Team$fromJSONString()}}
\item \href{#method-clone}{\code{Team$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$new(
  id,
  name,
  link,
  venue,
  abbreviation,
  triCode,
  teamName,
  locationName,
  firstYearOfPlay,
  division,
  conference,
  franchise,
  roster,
  nextGameSchedule,
  shortName,
  officialSiteUrl,
  franchiseId,
  active
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$fromJSON(TeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$fromJSONString(TeamJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Team$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
