% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameSchedule.r
\name{TeamNextGameSchedule}
\alias{TeamNextGameSchedule}
\title{TeamNextGameSchedule Class}
\description{
TeamNextGameSchedule Class

TeamNextGameSchedule Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{totalItems}}{}

\item{\code{totalEvents}}{}

\item{\code{totalGames}}{}

\item{\code{totalMatches}}{}

\item{\code{dates}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameSchedule$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameSchedule$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameSchedule$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameSchedule$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameSchedule$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameSchedule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$new(
  totalItems,
  totalEvents,
  totalGames,
  totalMatches,
  dates
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$fromJSON(TeamNextGameScheduleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$fromJSONString(TeamNextGameScheduleJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameSchedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
