% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameScheduleDates.r
\name{TeamNextGameScheduleDates}
\alias{TeamNextGameScheduleDates}
\title{TeamNextGameScheduleDates Class}
\description{
TeamNextGameScheduleDates Class

TeamNextGameScheduleDates Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{date}}{}

\item{\code{totalItems}}{}

\item{\code{totalEvents}}{}

\item{\code{totalGames}}{}

\item{\code{totalMatches}}{}

\item{\code{games}}{}

\item{\code{events}}{}

\item{\code{matches}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameScheduleDates$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameScheduleDates$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameScheduleDates$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameScheduleDates$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameScheduleDates$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameScheduleDates$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$new(
  date,
  totalItems,
  totalEvents,
  totalGames,
  totalMatches,
  games,
  events,
  matches
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$fromJSON(TeamNextGameScheduleDatesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$fromJSONString(TeamNextGameScheduleDatesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleDates$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
