% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameScheduleGames.r
\name{TeamNextGameScheduleGames}
\alias{TeamNextGameScheduleGames}
\title{TeamNextGameScheduleGames Class}
\description{
TeamNextGameScheduleGames Class

TeamNextGameScheduleGames Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gamePk}}{}

\item{\code{link}}{}

\item{\code{gameType}}{}

\item{\code{season}}{}

\item{\code{gameDate}}{}

\item{\code{status}}{}

\item{\code{teams}}{}

\item{\code{venue}}{}

\item{\code{content}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameScheduleGames$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameScheduleGames$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameScheduleGames$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameScheduleGames$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameScheduleGames$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameScheduleGames$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$new(
  gamePk,
  link,
  gameType,
  season,
  gameDate,
  status,
  teams,
  venue,
  content
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$fromJSON(TeamNextGameScheduleGamesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$fromJSONString(TeamNextGameScheduleGamesJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleGames$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
