% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameScheduleStatus.r
\name{TeamNextGameScheduleStatus}
\alias{TeamNextGameScheduleStatus}
\title{TeamNextGameScheduleStatus Class}
\description{
TeamNextGameScheduleStatus Class

TeamNextGameScheduleStatus Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{abstractGameState}}{}

\item{\code{codedGameState}}{}

\item{\code{detailedState}}{}

\item{\code{statusCode}}{}

\item{\code{startTimeTBD}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameScheduleStatus$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameScheduleStatus$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameScheduleStatus$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameScheduleStatus$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameScheduleStatus$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameScheduleStatus$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$new(
  abstractGameState,
  codedGameState,
  detailedState,
  statusCode,
  startTimeTBD
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$fromJSON(TeamNextGameScheduleStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$fromJSONString(TeamNextGameScheduleStatusJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleStatus$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
