% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameScheduleTeamsAway.r
\name{TeamNextGameScheduleTeamsAway}
\alias{TeamNextGameScheduleTeamsAway}
\title{TeamNextGameScheduleTeamsAway Class}
\description{
TeamNextGameScheduleTeamsAway Class

TeamNextGameScheduleTeamsAway Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{leagueRecord}}{}

\item{\code{score}}{}

\item{\code{team}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameScheduleTeamsAway$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameScheduleTeamsAway$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameScheduleTeamsAway$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameScheduleTeamsAway$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameScheduleTeamsAway$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameScheduleTeamsAway$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$new(leagueRecord, score, team)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$fromJSON(TeamNextGameScheduleTeamsAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$fromJSONString(TeamNextGameScheduleTeamsAwayJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsAway$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
