% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamNextGameScheduleTeamsHome.r
\name{TeamNextGameScheduleTeamsHome}
\alias{TeamNextGameScheduleTeamsHome}
\title{TeamNextGameScheduleTeamsHome Class}
\description{
TeamNextGameScheduleTeamsHome Class

TeamNextGameScheduleTeamsHome Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{leagueRecord}}{}

\item{\code{score}}{}

\item{\code{team}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamNextGameScheduleTeamsHome$new()}}
\item \href{#method-toJSON}{\code{TeamNextGameScheduleTeamsHome$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamNextGameScheduleTeamsHome$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamNextGameScheduleTeamsHome$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamNextGameScheduleTeamsHome$fromJSONString()}}
\item \href{#method-clone}{\code{TeamNextGameScheduleTeamsHome$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$new(leagueRecord, score, team)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$fromJSON(TeamNextGameScheduleTeamsHomeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$fromJSONString(TeamNextGameScheduleTeamsHomeJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamNextGameScheduleTeamsHome$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
