% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamsApi.r
\name{TeamsApi}
\alias{TeamsApi}
\title{Teams operations}
\description{
swagger.Teams
}
\section{Methods}{

\describe{

get_team Get an NHL team.


get_team_roster Get an NHL team's roster.


get_team_stats Get all statistics for an NHL team.


get_teams Get all NHL teams.

}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Stores url path of the request.}

\item{\code{apiClient}}{Handles the client-server communication.}

\item{\code{userAgent}}{Set the user agent of the request.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{path}}{Stores url path of the request.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamsApi$new()}}
\item \href{#method-get_team}{\code{TeamsApi$get_team()}}
\item \href{#method-get_team_roster}{\code{TeamsApi$get_team_roster()}}
\item \href{#method-get_team_stats}{\code{TeamsApi$get_team_stats()}}
\item \href{#method-get_teams}{\code{TeamsApi$get_teams()}}
\item \href{#method-clone}{\code{TeamsApi$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$new(apiClient)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team"></a>}}
\if{latex}{\out{\hypertarget{method-get_team}{}}}
\subsection{Method \code{get_team()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_team(id, expand, season, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_roster"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_roster}{}}}
\subsection{Method \code{get_team_roster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_team_roster(id, season, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_team_stats"></a>}}
\if{latex}{\out{\hypertarget{method-get_team_stats}{}}}
\subsection{Method \code{get_team_stats()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_team_stats(id, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_teams"></a>}}
\if{latex}{\out{\hypertarget{method-get_teams}{}}}
\subsection{Method \code{get_teams()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$get_teams(expand, season, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamsApi$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
