% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TeamStatsStat.r
\name{TeamStatsStat}
\alias{TeamStatsStat}
\title{TeamStatsStat Class}
\description{
TeamStatsStat Class

TeamStatsStat Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gamesPlayed}}{}

\item{\code{wins}}{}

\item{\code{losses}}{}

\item{\code{ot}}{}

\item{\code{pts}}{}

\item{\code{ptPctg}}{}

\item{\code{goalsPerGame}}{}

\item{\code{goalsAgainstPerGame}}{}

\item{\code{evGGARatio}}{}

\item{\code{powerPlayPercentage}}{}

\item{\code{powerPlayGoals}}{}

\item{\code{powerPlayGoalsAgainst}}{}

\item{\code{powerPlayOpportunities}}{}

\item{\code{penaltyKillPercentage}}{}

\item{\code{shotsPerGame}}{}

\item{\code{shotsAllowed}}{}

\item{\code{winScoreFirst}}{}

\item{\code{winOppScoreFirst}}{}

\item{\code{winLeadFirstPer}}{}

\item{\code{winLeadSecondPer}}{}

\item{\code{winOutshootOpp}}{}

\item{\code{winOutshotByOpp}}{}

\item{\code{faceOffsTaken}}{}

\item{\code{faceOffsWon}}{}

\item{\code{faceOffsLost}}{}

\item{\code{faceOffWinPercentage}}{}

\item{\code{shootingPctg}}{}

\item{\code{savePctg}}{}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{TeamStatsStat$new()}}
\item \href{#method-toJSON}{\code{TeamStatsStat$toJSON()}}
\item \href{#method-fromJSON}{\code{TeamStatsStat$fromJSON()}}
\item \href{#method-toJSONString}{\code{TeamStatsStat$toJSONString()}}
\item \href{#method-fromJSONString}{\code{TeamStatsStat$fromJSONString()}}
\item \href{#method-clone}{\code{TeamStatsStat$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$new(
  gamesPlayed,
  wins,
  losses,
  ot,
  pts,
  ptPctg,
  goalsPerGame,
  goalsAgainstPerGame,
  evGGARatio,
  powerPlayPercentage,
  powerPlayGoals,
  powerPlayGoalsAgainst,
  powerPlayOpportunities,
  penaltyKillPercentage,
  shotsPerGame,
  shotsAllowed,
  winScoreFirst,
  winOppScoreFirst,
  winLeadFirstPer,
  winLeadSecondPer,
  winOutshootOpp,
  winOutshotByOpp,
  faceOffsTaken,
  faceOffsWon,
  faceOffsLost,
  faceOffWinPercentage,
  shootingPctg,
  savePctg
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$fromJSON(TeamStatsStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$fromJSONString(TeamStatsStatJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TeamStatsStat$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
