% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplay-package.R, R/utils.R
\docType{package}
\name{powerplay-package}
\alias{powerplay}
\alias{powerplay-package}
\title{powerplay: Functions to Access Hockey Play by Play Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A utility to quickly obtain clean and tidy hockey play by play data.

\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A utility to quickly obtain clean and tidy hockey play by play data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saiemgilani/powerplay}
  \item \url{https://saiemgilani.github.io/powerplay/}
  \item Report bugs at \url{https://github.com/saiemgilani/powerplay/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/saiemgilani/powerplay}
  \item \url{https://saiemgilani.github.io/powerplay/}
  \item Report bugs at \url{https://github.com/saiemgilani/powerplay/issues}
}

}
\author{
\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com}


\strong{Maintainer}: Saiem Gilani \email{saiem.gilani@gmail.com}

}
\keyword{internal}
